netlink/requirements.txt
fastapi
uvicorn
mysql-connector-python
httpx
pytest
httpcore
pytest
pandas
matplotlib
pymysql
sqlalchemy
slowapi
python-dotenv
pydantic[email]
email-validator
logtail-python
cryptography
aioftp
minio
python-multipart
bcrypt
influxdb-client
black
flake8
isort
firebase-admin
asyncio
starlette
python-jose
